requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
aiogram==3.7.0
python-dotenv==1.0.1
pydantic==2.7.4
python-multipart==0.0.9
requests==2.32.3
